Update project prepare 3D surface), no more memory problem)
